Refactor callmeback into separate blueprint
