+++ cla_public/apps/callmeback/__init__.py
@@ -1,0 +1,5 @@
+"CallMeBack app"
+from flask import Blueprint
+
+
+callmeback = Blueprint('callmeback', __name__)
