+++ cla_public/apps/callmeback/fields.py
@@ -1,0 +1,182 @@
+# -*- coding: utf-8 -*-
+"CallMeBack form fields"
+
+import datetime
+
+from flask.ext.babel import lazy_gettext as _
+from wtforms import FormField, RadioField, SelectField
+from wtforms import Form as NoCsrfForm
+from wtforms.validators import InputRequired, ValidationError
+
+from cla_public.apps.callmeback.constants import DAY_CHOICES, DAY_TODAY, \
+    DAY_TOMORROW, DAY_SPECIFIC
+from cla_public.apps.checker.validators import IgnoreIf, FieldValueNot
+from cla_public.libs import call_centre_availability
+from cla_public.libs.call_centre_availability import available, \
+    available_days, day_choice, time_choice, time_slots, today_slots, \
+    tomorrow_slots
+
+
+class FormattedChoiceField(object):
+    """
+    Mixin for fields applying a formatting function to values
+    """
+
+    def process_data(self, value):
+        self.data = value
+        if value:
+            self.data = self._format(value)
+
+    def pre_validate(self, form):
+        choice_values = (v for v, _ in self.choices)
+        if self._format(self.data) not in choice_values:
+            raise ValueError(self.gettext('Not a valid choice'))
+
+
+class DayChoiceField(FormattedChoiceField, SelectField):
+    """
+    Select field with next `num_days` days as options
+    """
+
+    def __init__(self, num_days=6, *args, **kwargs):
+        super(DayChoiceField, self).__init__(*args, **kwargs)
+        self.choices = map(day_choice, available_days(num_days))
+
+    def process_formdata(self, valuelist):
+        if valuelist:
+            try:
+                year = int(valuelist[0][:4])
+                month = int(valuelist[0][4:6])
+                day = int(valuelist[0][6:])
+                self.data = datetime.date(year, month, day)
+            except ValueError:
+                self.data = None
+                raise ValueError(self.gettext('Not a valid date'))
+
+    def _format(self, value):
+        return '{:%Y%m%d}'.format(value)
+
+
+class TimeChoiceField(FormattedChoiceField, SelectField):
+    """
+    Select field with available time slots for a specific day as options
+    """
+
+    def __init__(self, choices_callback=None, validators=None, **kwargs):
+        super(TimeChoiceField, self).__init__(validators=validators, **kwargs)
+        self.choices = map(time_choice, choices_callback())
+
+    def process_formdata(self, valuelist):
+        if valuelist:
+            try:
+                hour = int(valuelist[0][:2])
+                minute = int(valuelist[0][2:])
+                self.data = datetime.time(hour, minute)
+            except ValueError:
+                self.data = None
+                raise ValueError(self.gettext('Not a valid time'))
+
+    def _format(self, value):
+        return '{:%H%M}'.format(value)
+
+
+class AvailableSlot(object):
+    """
+    Validates whether the selected time slot is available.
+    """
+
+    def __init__(self, day):
+        self.day = day
+
+    def __call__(self, form, field):
+        date = call_centre_availability.current_datetime()
+        if self.day == DAY_TOMORROW:
+            date = date + datetime.timedelta(days=1)
+        if self.day == DAY_SPECIFIC:
+            date = form.day.data
+        time = datetime.datetime.combine(date, field.data)
+        if not available(time):
+            raise ValidationError(
+                field.gettext(
+                    u"Can't schedule a callback at the requested time"))
+
+
+class AvailabilityCheckerForm(NoCsrfForm):
+    """
+    Subform allowing the user to select a time and date for a callback
+    """
+    specific_day = RadioField(
+        label=_(u'Arrange a callback time'),
+        choices=DAY_CHOICES,
+        default=DAY_TODAY)
+
+    # choices must be set dynamically as cache is not available at runtime
+    time_today = TimeChoiceField(
+        today_slots,
+        validators=[
+            IgnoreIf('specific_day', FieldValueNot(DAY_TODAY)),
+            AvailableSlot(DAY_TODAY)],
+        id='id_time_today')
+    time_tomorrow = TimeChoiceField(
+        tomorrow_slots,
+        validators=[
+            IgnoreIf('specific_day', FieldValueNot(DAY_TOMORROW)),
+            AvailableSlot(DAY_TOMORROW)],
+        id='id_time_tomorrow')
+    day = DayChoiceField(
+        validators=[
+            IgnoreIf('specific_day', FieldValueNot(DAY_SPECIFIC)),
+            InputRequired()],
+        id='id_day')
+    time_in_day = TimeChoiceField(
+        time_slots,
+        validators=[
+            IgnoreIf('specific_day', FieldValueNot(DAY_SPECIFIC)),
+            AvailableSlot(DAY_SPECIFIC)],
+        id='id_time_in_day')
+
+    def __init__(self, *args, **kwargs):
+        kwargs['prefix'] = ''
+        super(AvailabilityCheckerForm, self).__init__(*args, **kwargs)
+        if not self.time_today.choices:
+            self.specific_day.data = DAY_TOMORROW
+        if not self.time_tomorrow.choices and \
+                self.specific_day.data == DAY_TOMORROW:
+            self.specific_day.data = DAY_SPECIFIC
+
+    def scheduled_time(self, today=None):
+        """
+        Get the datetime of the selected day and timeslot
+        """
+        date = today or datetime.date.today()
+        time = None
+
+        if self.specific_day.data == DAY_TODAY:
+            time = self.time_today.data
+
+        if self.specific_day.data == DAY_TOMORROW:
+            date += datetime.timedelta(days=1)
+            time = self.time_tomorrow.data
+
+        if self.specific_day.data == DAY_SPECIFIC:
+            date = self.day.data
+            time = self.time_in_day.data
+
+        if date and time:
+            return datetime.datetime.combine(date, time)
+
+        return None
+
+
+class AvailabilityCheckerField(FormField):
+    """Convenience class for FormField(AvailabilityCheckerForm"""
+
+    def __init__(self, *args, **kwargs):
+        super(AvailabilityCheckerField, self).__init__(
+            AvailabilityCheckerForm, *args, **kwargs)
+
+    def scheduled_time(self):
+        """
+        Get the datetime of the selected day and timeslot
+        """
+        return self.form.scheduled_time()
