+++ cla_public/apps/checker/views.py
@@ -10,15 +10,15 @@ from flask import abort, render_template, redirect, \
 from cla_public.apps.checker import checker
 from cla_public.apps.checker.api import post_to_case_api, \
     post_to_eligibility_check_api, get_organisation_list
-from cla_public.apps.checker.constants import RESULT_OPTIONS, CATEGORIES, ORGANISATION_CATEGORY_MAPPING, \
-    NO_CALLBACK_CATEGORIES
-from cla_public.apps.checker.decorators import form_view, override_session_vars, redirect_if_no_session, \
-    redirect_if_ineligible
+from cla_public.apps.callmeback.forms import CallMeBackForm
+from cla_public.apps.checker.constants import RESULT_OPTIONS, CATEGORIES, \
+    ORGANISATION_CATEGORY_MAPPING, NO_CALLBACK_CATEGORIES
+from cla_public.apps.checker.decorators import form_view, \
+    redirect_if_no_session, redirect_if_ineligible
 from cla_public.apps.checker.forms import AboutYouForm, YourBenefitsForm, \
     ProblemForm, PropertiesForm, SavingsForm, TaxCreditsForm, income_form, \
-    OutgoingsForm, ApplicationForm
-from cla_public.apps.checker.honeypot import FIELD_NAME as HONEYPOT_FIELD_NAME
-from cla_public.libs.utils import override_locale
+    OutgoingsForm
+from cla_public.libs.honeypot import FIELD_NAME as HONEYPOT_FIELD_NAME
 
 
 log = logging.getLogger(__name__)
@@ -32,7 +32,9 @@ def outcome(outcome):
     return proceed('result', outcome=outcome)
 
 
-checker.add_app_template_global(HONEYPOT_FIELD_NAME, name='honeypot_field_name')
+checker.add_app_template_global(
+    HONEYPOT_FIELD_NAME,
+    name='honeypot_field_name')
 
 
 @checker.after_request
@@ -188,15 +190,13 @@ def result(outcome):
         session.clear()
         return render_template('result/eligible-no-callback.html')
 
-    form = ApplicationForm()
+    form = CallMeBackForm()
     if form.validate_on_submit():
         if form.extra_notes.data:
             session.add_note(u'User problem:\n{0}'.format(form.extra_notes.data))
 
         post_to_eligibility_check_api(session.notes_object())
         post_to_case_api(form)
-
-        session['time_to_callback'] = form.time.scheduled_time()
 
         return redirect(url_for('.result', outcome='confirmation'))
 
@@ -211,13 +211,6 @@ def result(outcome):
         session.clear()
 
     return response
-
-
-@checker.route('/call-me-back', methods=['GET', 'POST'])
-@redirect_if_no_session()
-@form_view(ApplicationForm, 'call-me-back.html')
-def call_me_back(user):
-    return outcome('confirmation')
 
 
 @checker.route('/help-organisations/<category_name>', methods=['GET'])
