+++ cla_public/templates/index.html
@@ -12,7 +12,7 @@
 
   <p>{{ _('If CLA can’t help, you’ll get suggestions of other places you could go for advice.') }}</p>
 
-  <p><a class="button button-get-started" href="{{ url_for('checker.problem') }}" id="start">{{ _('Start') }}</a></p>
+  <p><a class="button button-get-started" href="{{ url_for('base.get_started') }}" id="start">{{ _('Start') }}</a></p>
 
   <h3>{{ _('Before you start') }}</h3>
   <p>{{ _('To use this service, you’ll need details of the following, for you and your partner, if you have one') }}:</p>
