+++ cla_public/templates/macros/element.html
@@ -34,7 +34,7 @@
   {% call alert() %}
     <p>
       {{ _('If you’re having trouble completing this form, ask us to') }}
-      <a href="{{ url_for('checker.call_me_back') }}" id="callback-link">{{ _('call you back') }}</a>.
+      <a href="{{ url_for('callmeback.request_callback') }}" id="callback-link">{{ _('call you back') }}</a>.
     </p>
   {% endcall %}
 {% endmacro %}
