+++ cla_public/templates/problem.html
@@ -31,7 +31,7 @@
       <p>{{ _('If you have more than one problem, you can complete this service more than once.') }}</p>
       <p>
         {{ _('If you’re really not sure what to choose, ask us to') }}
-        <a href="{{ url_for('.call_me_back') }}" id="callback-link">{{ _('call you back') }}</a>.
+        <a href="{{ url_for('callmeback.request_callback') }}" id="callback-link">{{ _('call you back') }}</a>.
       </p>
     {% endcall %}
